Updated functions.py
I streamlined the code for update_force and initial_conditions(). The force vector doesn't need to be in the initial condition with zero values, as we determine the force vector purely from position and velocity, not the previous force vector. 

So I had the force vector initialized in update_force (should probably change that to find_force, cause that's what it does). Also made the function not require inputting the force vector, considering it doesn't do anything with that input beyond overwriting it.