Update main_loop.py
I changed Main_loop.py to be in line with the initial conditions in terms of the force vector stuff from the previous commit. So we weren't using the force vector as a initial parameter just to use the function when it doesn't need to be there. 

But I made some other changes: I added in a history for the velocity so we could make velocity graphs if we wanted to. Then I also updated the history so it was actually recording the initial conditions in those lists, and not only the updated ones.